Move custom auth out of grants to auth
